refactor: move "TreeToObectMapper" to "compose/" folder
